+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## version 1.1.0, release date 2022-03-06
+
+- Added guest mode for anyone to use
+- Added mobile support
+- Added hotlink bypass login with /#/login?channel=XXXX
+
 ## version 1.0.0, release date 2022-02-22
 
 - First release